# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: PaiiD Trading API
  description: |
    Personal Artificial Intelligence Investment Dashboard

    AI-powered trading platform with real-time market data and intelligent trade execution.

    ## Data Sources
    - **Market Data:** Tradier API (Real-time, NO delay)
    - **Trade Execution:** Alpaca Paper Trading API
    - **AI Analysis:** Anthropic Claude API

    ## Authentication
    All endpoints (except health checks and public endpoints) require JWT Bearer token authentication.
    State-changing operations (POST/PUT/DELETE/PATCH) also require CSRF token in X-CSRF-Token header.
  version: 1.0.0
  contact:
    name: PaiiD Support
    url: https://github.com/your-repo/paiid
    email: support@paiid.com
  license:
    name: Proprietary
    url: https://paiid.com/terms
servers:
  - url: https://paiid-backend.onrender.com
    description: Production server (Render)
  - url: http://localhost:8001
    description: Local development server
security:
  - bearerAuth: []
  - csrfToken: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login/register endpoint
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token for state-changing operations (POST/PUT/DELETE/PATCH)
  schemas:
...
    Quote:
      type: object
      properties:
        symbol:
          type: string
          example: AAPL
        bid:
          type: number
          format: float
          example: 154.25
        ask:
          type: number
          format: float
          example: 154.50
        last:
          type: number
          format: float
          example: 154.35
        volume:
          type: integer
          example: 45678910
        timestamp:
          type: string
          format: date-time
          example: 2025-10-26T14:30:00.000Z
        cached:
          type: boolean
          example: false
    Bar:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: 2025-10-26T00:00:00Z
        open:
          type: number
          format: float
          example: 154.00
        high:
          type: number
          format: float
          example: 156.50
        low:
          type: number
          format: float
          example: 153.25
        close:
          type: number
          format: float
          example: 155.75
        volume:
          type: integer
          example: 78954210

    # Position Models
...
    Error:
      type: object
      properties:
        detail:
          type: string
          example: Authentication failed
...
  responses:
...
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Resource not found
...
paths:
...
  /api/market/quote/{symbol}:
    get:
      tags: [market-data]
      summary: Get real-time quote
      description: Get real-time quote for a symbol using Tradier (cached with configurable TTL)
      operationId: getQuote
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 10
          description: Stock symbol
          example: AAPL
      responses:
        '200':
          description: Quote data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/market/quotes:
    get:
      tags: [market-data]
      summary: Get multiple quotes
      description: Get quotes for multiple symbols (batch request)
      operationId: getQuotes
      parameters:
        - name: symbols
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
          description: Comma-separated symbols (e.g., "AAPL,MSFT,GOOGL")
          example: AAPL,MSFT,GOOGL
      responses:
        '200':
          description: Quotes data
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/Quote'
  /api/market/bars/{symbol}:
    get:
      tags: [market-data]
      summary: Get historical bars
      description: Get historical price bars (OHLCV data)
      operationId: getBars
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
          example: AAPL
        - name: timeframe
          in: query
          required: false
          schema:
            type: string
            enum: [1Min, 5Min, 15Min, 1Hour, 1Day, daily, weekly, monthly]
            default: daily
          example: daily
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          example: 100
      responses:
        '200':
          description: Historical bars
          content:
            application/json:
              schema:
                type: object
                properties:
                  symbol:
                    type: string
                    example: AAPL
                  bars:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bar'
                  cached:
                    type: boolean
                    example: false

  # Portfolio Endpoints
...